+++ packages/embed-wallet/src/EmbedWallet.ts
@@ -34,7 +34,12 @@ const buildEnvMap: Record<WalletEnvironment, TORUS_BUILD_ENV_TYPE> = {
   local: 'development',
   dev: 'cere-dev',
   stage: 'cere-stage',
-  prod: 'cere',
+  /**
+   * TODO: Change back to 'cere' when the MetaMask block issue is resolved
+   *
+   * @see https://github.com/MetaMask/eth-phishing-detect/issues/19072
+   */
+  prod: 'cere-aws',
 };
 
 const createBalance = (