+++ src/stores/ApprovalStore/convertPrice.ts
@@ -1,6 +1,6 @@
-import { BigNumber } from 'ethers';
+import { BigNumber, utils } from 'ethers';
 import type { TokenConfig } from '@cere-wallet/wallet-engine';
 
 export const convertPrice = (amount: BigNumber, { decimals }: TokenConfig) => {
-  return amount.div(10 ** decimals).toNumber();
+  return +utils.formatUnits(amount, decimals);
 };