+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 - 
 
+### v1.31.0
+
+- Switch to new Freeport Smart Contract SDK
+- Refactor Assets and activity stores to reduce copy & past
+- Remove usages of static token config everywhere, and fetch the data from contracts
+- Fix custom assets activity capturing and decimals detection
+
 ### v1.30.0
 
 - Add Wallet Permission System ([EIP-2255](https://eips.ethereum.org/EIPS/eip-2255)) support